Committed by Luca Cerutti 02 may 2013.
- added a VC11 solution
- code for dllfreeWRL was moved under \src\dllFreeWRL so that each
project can point to it
- code for np_freeX3D was moved under \src\plugin_win32 so that each
project can point to it
- code for freeWRLEAItest is now only under the VC9_projectfiles folder.
Since it is a test application I believed that was not right to set it
under \src

- src\libeai\EAI_c_node.c modified for handling field definition retrieval
- src\libeai\X3D_node.h modified for handling field definition retrieval

- \projectfiles_vc9\freeWRLEAITest\Ctest_libeai_Teardown.c added test
methods for field definitions retrieval (they need more refining)

- \src\dllFreeWRL\dllFreeWRL.h/cpp    changed in order to support "eai"
parameter for use with eai
- \src\plugin_win32\ changed code along the initialization pipeline in
order to read the "eai" flag.
- \src\plugin_win32\test.html changed to show the use of the "eai"
params. The code is commented, since the firewall will kick in (at least
as long we use EAI on sockets) and ask permission