fixed build issues on src/java
Paths were incorrect (builddir instead of srcdir) for manifest.extras, causing
out-of-source builds to fail.  Fixed.

One .java file may compile to multiple .class files, but the extra .class files
were not being included in the jar; fixed this by changing JAVAROOT to a custom
builddir, and include all contents of this dir in the jar

Signed-off-by: Ian Stakenvicius <axs@gentoo.org>