autotools build system: vtempl script is built, fixed refpaths
The vtempl script is called to help with version tagging when using the autotools based
build system.  It is "built" from vtempl.in, and so it is created in the builddir.  When
doing an in-source build this doesn't much matter, but an out-of-source build needs the
path to be specified properly.  This commit changes the path to $(top_builddir) instead
of $(top_srcdir).