autotools build system: fix standard-statusbar codepath
ui/statusbar.c, used when --with-statusbar=standard is set, requires opengl/RasterFont.{c,h}
to be included in the project and some code in the globals object that was commented out.
This commit adds RasterFont.h to Makefile.sources, RasterFont.c alongside statusbar.c in Makefile.am,
and an '#ifdef STATUSBAR_STD'-wrapped inclusion of the commented-out code in iglobals.  Results
were compile-tested on linux and a 'make dist' cleanly includes all code files now.